+++ src/engine/Collision/Detection/DynamicTreeCollisionProcessor.ts
@@ -17,6 +17,8 @@ import { RayCastHit } from './RayCastHit';
 import { DeepRequired } from '../../Util/Required';
 import { PhysicsConfig } from '../PhysicsConfig';
 import { RayCastOptions } from './RayCastOptions';
+import { BoundingBox } from '../BoundingBox';
+import { createId } from '../../Id';
 
 /**
  * Responsible for performing the collision broadphase (locating potential collisions) and
@@ -37,6 +39,38 @@ export class DynamicTreeCollisionProcessor implements CollisionProcessor {
 
   public getColliders(): readonly Collider[] {
     return this._colliders;
+  }
+
+  public query(point: Vector): Collider[];
+  public query(bounds: BoundingBox): Collider[];
+  public query(pointOrBounds: Vector | BoundingBox): Collider[] {
+    const results: Collider[] = [];
+    if (pointOrBounds instanceof BoundingBox) {
+      this._dynamicCollisionTree.query(
+        {
+          id: createId('collider', -1),
+          owner: null,
+          bounds: pointOrBounds
+        } as Collider,
+        (other) => {
+          results.push(other);
+          return false;
+        }
+      );
+    } else {
+      this._dynamicCollisionTree.query(
+        {
+          id: createId('collider', -1),
+          owner: null,
+          bounds: new BoundingBox(pointOrBounds.x, pointOrBounds.y, pointOrBounds.x, pointOrBounds.y)
+        } as Collider,
+        (other) => {
+          results.push(other);
+          return false;
+        }
+      );
+    }
+    return results;
   }
 
   public rayCast(ray: Ray, options?: RayCastOptions): RayCastHit[] {